+++ src/main/java/nextstep/auth/authentication/TokenAuthenticationInterceptor.java
@@ -7,45 +7,20 @@ import javax.servlet.http.HttpServletResponse;
 import nextstep.auth.context.Authentication;
 import nextstep.auth.token.JwtTokenProvider;
 import nextstep.auth.token.TokenResponse;
-import nextstep.member.application.CustomUserDetailsService;
-import nextstep.member.domain.LoginMember;
+import nextstep.member.application.UserDetailService;
 import org.springframework.http.MediaType;
 
 public class TokenAuthenticationInterceptor extends AuthenticationInterceptor {
 
-    private final AuthenticationConverter authenticationConverter;
-    private final CustomUserDetailsService userDetailsService;
     private final JwtTokenProvider jwtTokenProvider;
 
 
-    public TokenAuthenticationInterceptor(CustomUserDetailsService userDetailsService,
+    public TokenAuthenticationInterceptor(UserDetailService userDetailsService,
         JwtTokenProvider jwtTokenProvider,
         AuthenticationConverter authenticationConverter) {
-        this.userDetailsService = userDetailsService;
+        super(userDetailsService, authenticationConverter);
+
         this.jwtTokenProvider = jwtTokenProvider;
-        this.authenticationConverter = authenticationConverter;
-    }
-
-    @Override
-    public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) throws IOException {
-        AuthenticationToken token = convert(request);
-        Authentication authentication = authenticate(token);
-        afterAuthentication(request, response, authentication);
-
-        return false;
-    }
-
-    public AuthenticationToken convert(HttpServletRequest request) {
-        return authenticationConverter.convert(request);
-    }
-
-    public Authentication authenticate(AuthenticationToken token) {
-        // TODO: AuthenticationToken에서 Authentication 객체 생성하기
-        String principal = token.getPrincipal();
-        LoginMember loginMember = userDetailsService.loadUserByUsername(principal);
-        checkAuthentication(loginMember, token);
-
-        return new Authentication(loginMember);
     }
 
     @Override
