+++ src/main/java/nextstep/member/application/UserDetailService.java
@@ -1,0 +1,9 @@
+package nextstep.member.application;
+
+import nextstep.member.domain.LoginMember;
+
+public interface UserDetailService {
+
+    LoginMember loadUserByUsername(String email);
+
+}
